shot code move to androidDemo App.
