+++ androidDemoApp/build.gradle.kts
@@ -7,7 +7,7 @@ plugins {
     alias(libs.plugins.kotlinx.serialization)
     alias(libs.plugins.gradle.play.publisher)
 }
-
+apply(plugin = "shot")
 apply(from = rootProject.file("gradle/version.gradle.kts"))
 
 fun calculateVersionCode(): Int = (extra["calculateVersionCode"] as () -> Int).invoke()
@@ -34,8 +34,7 @@ android {
         applicationId = "com.michaeltchuang.walletsdk.demo"
         versionCode = calculateVersionCode()
         versionName = calculateVersionName()
-
-        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner = "com.karumi.shot.ShotTestRunner"
     }
 
     buildFeatures {
@@ -172,6 +171,7 @@ dependencies {
     implementation(project(":sharedDemoApp"))
     implementation(project(":wallet-sdk-ui"))
     implementation(project(":wallet-sdk-core"))
+    implementation(libs.androidx.material3)
 
     implementation(project.dependencies.platform(libs.koin.bom))
     implementation(libs.androidx.activityCompose)
@@ -188,4 +188,12 @@ dependencies {
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.uiautomator)
     debugImplementation(libs.compose.ui.testManifest)
+
+    androidTestImplementation(libs.compose.ui.tooling)
+    androidTestImplementation(libs.androidx.compose.material3)
+    androidTestImplementation(libs.androidx.compose.foundation)
+    androidTestImplementation(libs.androidx.compose.ui)
+    androidTestImplementation(libs.navigation.compose)
+
+    androidTestImplementation(libs.junit)
 }
