+++ androidDemoApp/src/androidTest/kotlin/com/michaeltchuang/walletsdk/ui/accountdetails/screens/ShowAddressScreenshotTest.kt
