+++ androidDemoApp/src/androidTest/kotlin/com/michaeltchuang/walletsdk/ui/accountdetails/screens/ViewPassphraseScreenshotTest.kt
