+++ androidDemoApp/src/androidTest/kotlin/com/michaeltchuang/walletsdk/ui/liquidAuth/screens/LiquidAuthScreenshotTest.kt
