+++ androidDemoApp/src/androidTest/kotlin/com/michaeltchuang/walletsdk/ui/onboarding/screens/Falcon24WalletSelectionScreenshotTest.kt
