+++ androidDemoApp/src/androidTest/kotlin/com/michaeltchuang/walletsdk/ui/onboarding/screens/RecoveryPhraseScreenshotTest.kt
