+++ androidDemoApp/src/androidTest/kotlin/com/michaeltchuang/walletsdk/ui/qrscanner/screens/QRScannerScreenshotTest.kt
