+++ androidDemoApp/src/androidTest/kotlin/com/michaeltchuang/walletsdk/ui/settings/screens/HDWalletSelectionScreenshotTest.kt
