+++ androidDemoApp/src/androidTest/kotlin/com/michaeltchuang/walletsdk/ui/settings/screens/LanguageSelectorScreenshotTest.kt
