+++ androidDemoApp/src/androidTest/kotlin/com/michaeltchuang/walletsdk/ui/settings/screens/NodeSettingsScreenshotTest.kt
