+++ androidDemoApp/src/androidTest/kotlin/com/michaeltchuang/walletsdk/ui/signing/screens/AddAssetScreenshotTest.kt
