+++ androidDemoApp/src/androidTest/kotlin/com/michaeltchuang/walletsdk/ui/signing/screens/AssetTransferConfirmScreenshotTest.kt
