+++ androidDemoApp/src/androidTest/kotlin/com/michaeltchuang/walletsdk/ui/signing/screens/SelectAccountScreenshotTest.kt
