+++ androidDemoApp/src/androidTest/kotlin/com/michaeltchuang/walletsdk/ui/signing/screens/SelectReceiverScreenshotTest.kt
