+++ gradle/libs.versions.toml
@@ -90,6 +90,7 @@ stream-webrtc = "1.3.10"
 uiautomator = "2.4.0"
 uuid-generator = "5.2.0"
 webviewMultiplatform = "1.3.0"
+material3 = "1.4.0"
 
 [libraries]
 
@@ -244,6 +245,7 @@ jackson-dataformat-msgpack = { module = "org.msgpack:jackson-dataformat-msgpack"
 jackson-dataformat-cbor = { module = "com.fasterxml.jackson.dataformat:jackson-dataformat-cbor", version.ref = "jackson-dataformat-cbor" }
 json-kotlin-schema = { module = "net.pwall.json:json-kotlin-schema", version.ref = "json-kotlin-schema" }
 json-canonicalization = { module = "io.github.erdtman:java-json-canonicalization", version.ref = "jsonCanonicalization" }
+androidx-material3 = { group = "androidx.compose.material3", name = "material3", version.ref = "material3" }
 
 [plugins]
 android-application = { id = "com.android.application", version.ref = "agp" }
