+++ wallet-sdk-ui/src/commonMain/kotlin/com/michaeltchuang/walletsdk/ui/accountdetails/screens/AccountDetailScreen.kt
@@ -114,7 +114,7 @@ fun AccountDetailScreen(
 }
 
 @Composable
-internal fun ScreenContent(
+ fun ScreenContent(
     navController: NavController,
     address: String,
     viewState: AccountDetailViewModel.ViewState,
