+++ wallet-sdk-ui/src/commonMain/kotlin/com/michaeltchuang/walletsdk/ui/onboarding/screens/OnboardingAccountTypeScreen.kt
@@ -123,7 +123,7 @@ fun OnboardingAccountTypeScreen(
 }
 
 @Composable
-internal fun ScreenContent(
+ fun ScreenContent(
     hasWalletWithNoAccounts: Boolean,
     isLoading: Boolean = false,
     onCreateNewAccount: () -> Unit = {},
