+++ wallet-sdk-ui/src/commonMain/kotlin/com/michaeltchuang/walletsdk/ui/onboarding/screens/RecoveryPhraseScreen.kt
@@ -155,7 +155,7 @@ fun RecoveryPhraseScreen(
 }
 
 @Composable
-internal fun ScreenContent(
+ fun ScreenContent(
     navController: NavController,
     accountType: AccountMnemonic.AccountType,
     mnemonicList: List<String>,
