+++ wallet-sdk-ui/src/commonMain/kotlin/com/michaeltchuang/walletsdk/ui/settings/screens/DeveloperSettingsScreen.kt
@@ -62,7 +62,7 @@ fun DeveloperSettingsScreen(
 }
 
 @Composable
-internal fun ScreenContent(
+ fun ScreenContent(
     navController: NavController,
     onCreateAlgoAccount: () -> Unit = {},
 ) {
