+++ wallet-sdk-ui/src/commonMain/kotlin/com/michaeltchuang/walletsdk/ui/settings/screens/LanguageSelectorScreen.kt
@@ -66,7 +66,7 @@ fun LanguageScreen(navController: NavController) {
 }
 
 @Composable
-internal fun ScreenContent(
+ fun ScreenContent(
     navController: NavController,
     viewState: LanguageSelectorViewModel.ViewState,
     onLanguageSelected: (LocalizationPreference) -> Unit = {},
