+++ wallet-sdk-ui/src/commonMain/kotlin/com/michaeltchuang/walletsdk/ui/settings/screens/NodeSettingsScreen.kt
@@ -95,7 +95,7 @@ fun NodeSettingsScreen(navController: NavController) {
 }
 
 @Composable
-internal fun ScreenContent(
+ fun ScreenContent(
     navController: NavController,
     viewState: NodeSettingsViewModel.ViewState,
     showMainnetWarningDialog: Boolean = false,
