+++ wallet-sdk-ui/src/commonMain/kotlin/com/michaeltchuang/walletsdk/ui/signing/screens/AddAssetScreen.kt
@@ -125,7 +125,7 @@ fun AddAssetScreen(
 }
 
 @Composable
-internal fun ScreenContent(
+ fun ScreenContent(
     navController: NavController,
     viewState: AddAssetViewModel.ViewState,
     snackbarHostState: SnackbarHostState = remember { SnackbarHostState() },
