+++ wallet-sdk-ui/src/commonMain/kotlin/com/michaeltchuang/walletsdk/ui/signing/screens/KeyReyConfirmScreen.kt
@@ -99,7 +99,7 @@ fun ConfirmTransactionRequestScreen(
 }
 
 @Composable
-internal fun ScreenContent(
+ fun ScreenContent(
     viewState: KeyRegConfirmViewModel.ViewState,
     onConfirm: () -> Unit,
     onBack: () -> Unit,
