+++ wallet-sdk-ui/src/commonMain/kotlin/com/michaeltchuang/walletsdk/ui/signing/screens/SelectReceiverScreen.kt
@@ -118,7 +118,7 @@ fun SelectReceiverScreen(
 }
 
 @Composable
-internal fun ScreenContent(
+ fun ScreenContent(
     navController: NavController,
     viewState: SelectReceiverViewModel.ViewState,
     onSearchTextChange: (String) -> Unit = {},
